Nintendo 3ds(armv6) Got a PoC up and running

+++ libs/base/sources/mem.c
@@ -48,7 +48,7 @@ void *memset(void *dest, const uint8_t val, size_t len)
 // Returns 1 if src is less than dst
 // Returns -1 if src is greater than dst
 // Returns 0 if the length of src and dst are equal
-uint64_t memcmp(const void *src, void *dst, size_t n)
+uint64_t memcmp(const void *src, void *dst, uint64_t n)
 {
 	const uint8_t *src_a = src;
 	uint8_t *dst_a = dst;
